requirements.txt
backports.shutil-get-terminal-size==1.0.0
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
Django==1.11.13
django-rest-framework==0.1.0
djangorestframework==3.8.2
enum34==1.1.6
idna==2.6
ipython==5.7.0
ipython-genutils==0.2.0
Markdown==2.6.11
pathlib2==2.3.2
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2018.4
requests==2.18.4
scandir==1.7
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
